scripts/gpu_test: Fixed index bug and revamped thread killing logic.
This fixes the linked bug which is about a simple off-by-one error, but
also refactors the way we invoke and keep track of glxgears
windows/threads. Instead of relying on the window ID, which may not be
found, we keep track of the thread's child process id, and the terminate
method uses that to reliably kill any spawned process.

All this should eliminate cases where a rogue thread keeps the entire
script from exiting thus hanging the test run.

We also handle the (corner) case where if glxgears fails to run, the
thread will just exit with a harmless message and no nasty trace.