plainbox:providers: fix scripts/bin duality
After the recent provider reorganization executable scripts were
supposed to be located in the 'bin' directory relative to the base
location of each provider.

There are two exceptions though: The CheckBoxSrcProvider is using
whatever is really in the source tree (and for legacy support that it
still 'scripts'). To fix this is now mixes with the new
ExecutablesInScriptsMixIn class. The StubBoxProvider has no legacy
constraints but is still holding the single executable it has in the
'scripts' directory so to make it work it also mixes with the new helper
class.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>