plainbox: replace JobDefinition.uses_policykit with .secure
The uses_policykit attribute, which was added recently when creating the
multi-provider support code, as a way to let applications know if they
should "warm up" pkexec prompt for the checkbox-trusted-launcher.

Since the name (and how it was derived) was found to be inadequate
I'd like to rename the property to 'secure' to indicate that the trusted
launcher can discover and run this job.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>