plainbox:secure:launcher1: add plainbox-trusted-launcher-1
This patch replaces checkbox-trusted-launcher (which was broken since we
started the process of introducing support for multiple providers) with
plainbox-trusted-launcher-1.

The new launcher works pretty much the same way as the old one did, with
one exception: it supports all of the v1 providers loaded from
system-wide directory. It is thus, no longer, checkbox specific in any
way.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>