move the sru command to checkbox-ng
The `plainbox sru` command was historically implemented underneath the
plainbox executable since that was the only thing we had. This collides
with the intent to make plainbox a low-level development tool and a
framework for checkbox (and other tools). It was somewhat legitimized by
the daily sru testing using "plainbox" as opposed to checkbox.

With the current focus to release plainbox to Ubuntu and to do a proper
upstream release of plainbox source I'd like to move this very
business-specific command to checkbox-ng.

This patch moves the raw command implementation, adjusts some tests and
documentation, where sru was being mentioned.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>