rebrand 'plainbox ci' to 'checkbox sru ci'
This patch moves and changes three files from plainbox/plainbox/data/ to
checkbox-ng/contrib. Those files are picked up by the debian packaging
to implement the SRU CI loop that runs every day.

This is mostly an automatic change with one exception. The
externally-generated CI config file is still called (and read as)
plainbox, for compatibility.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>