plainbox:ctrl: use env trick with sudo controller
This patch corrects one of the many faces of the 'path bug' (executables
not being in PATH in certain situations).

One aspect of that bug is how that 'sudo -E' which was being used in the
sudo execution controller is especially sensitive to ignore PATH even if
-E is used. Instead it loads the "secure path" from system wide
configuration file. This feature would obviously erase our customized
PATH and thus jobs using sudo this way would fail to see nest_dir with
all the executables.

The fix is to bring the sudo controller in alignment with pkexec
controller that already uses env(1) to pass arbitrary environment
variables into a program executed with sudo (without the -E flag this
time).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>