plainbox: fix tests to run winthout ../checkbox-old
This patch changes some of the unit tests to either switch away from
using checkbox-old as a job provider (to stubbox) or to gracefuly fail
(skip tests) if checkbox-old is unavailable. The latter case is going to
be migrated away to per-provider tests so I didn't want to completely
remove it until that transition (provider tests) is completed.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>