checkbox_ng:commands:cli: Use the curses functions only when running in a tty.
Otherwise the default whitelist is automatically selected and the command
will skip the welcome text.
CLi commands will exit if no whitelists are selected.

Whitelists displayed are those belonging to the list of default providers.

A common problem when debugging a curses application is to get your terminal
messed up when the application dies without restoring the terminal to its
previous state. Keys are no longer be echoed to the screen when you type
them, for example, which makes using the shell difficult.

In Python you can avoid these complications and make debugging much easier
by importing the module curses.wrapper. It supplies a wrapper() function
that takes a callable. It does the curses init and also initializes colors
if color support is present. It then runs the provided callable and finally
deinitializes appropriately. The callable is called inside a try-catch
clause which catches exceptions, performs curses deinitialization, and then
passes the exception upwards. Thus, the terminal won’t be left in a funny
state on exception.

Tested with CDTS and checkbox-certification welcome texts, both render
correctly in 80x24 terminals.

Curses raises an exception if it cannot write to a stdscr position.
It happens if terminal size is below Ubuntu default values.
This type of curses.errors are caught and checkbox-ng exits with:
SystemExit('Terminal size must be at least 80x24')