plainbox:commands:test_run: Mock concurrent modules imported by highlevel
test_run calls box.py main function and doing so imports the dev command
which also imports the list command.
The list command imports highlevel and this module does:

from concurrent.futures import ThreadPoolExecutor

That leads the concurrent.futures.thread module to be imported, and with it
an exit function to be registered:

atexit.register(_python_exit)

The problem seems to be that concurrent.futures which is loaded in by
unittest is then unloaded before the _exit_function gets called, and it's
function definitions by the way.