plainbox:commands: rewrite with plainbox.impl.clitools
This patch rewrites parts of the PlainBoxCommand and PlainBoxToolBase
on top of the equivalent classes from the clitools module.

The only notable change is how late_init() sets the _provider_list
instance attribute, it used to be re-set (overwritten) but now we extend
the list of providers instead. This is because the same list has already
been passed to subcommand initializers in add_subcommands().

The method that computes the list of providers has been slightly
simplified, by dropping the 'ihv' use case and making it the same as
'auto'.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>