Replaces ubuntu-sdk requisite with expanded dependency list for tarmac/vagrant testing.
ubuntu-sdk pulls in a series of packages not needed for command-line
building; further, some of these packages sometimes have bad
dependencies themselves, rendering ubuntu-sdk uninstallable which
hinders building and testing of our packages.

We looked at ubuntu-sdk's dependencies, culled the bad ones and a few
that we are certain we don't require, and replaced the blanket
ubuntu-sdk dependency with this extended list. This should give us more
control and feedback and as a bonus tests and builds should be slightly
faster.