qml:RunManagerTestDelegate: Replace the Rerun icon by a standard Switch
MouseArea onClicked and onRerunStatusChanged are merged into a single signal
handler onCheckedChanged to avoid managing a tri-state element.

Define a rerunCount variable in RunManagerView to only disable the rerun
button if no tests are selected for rerun. The testListModel cannot be used
for this purpose as local jobs can also be implicitly selected.

If a test is skipped because of dependencies not met , both the rerun Switch
and the log button are disabled.

A new runstatus value (99) is used to distinguish test selected for rerun
from tests that have not run yet. That way the status of already run tests
are not masked when they are selected for rerun, useful if the tester wants
to only rerun failed tests, the failed marks are still displayed.

Adjust margins since the Switch width is twice the size of a regular icon.