plainbox: make FsPlugInCollection deal with list of directories
This patch breaks the FsPlugInCollection API by changing the first
argument from a string consisting of of os.path.pathsep-separated
directories to a list of directories.

This makes it possible to use the path separator, ':' on unix platforms,
in the directory itself. This fixes a bug where the typical provider
name would contain a colon and thus break job and whitelist discovery.
By accident, the existing providers are not using colons in their name
yet so this bug was not manifesting itself unless one wanted to work
with a custom provider.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>