plainbox:secure:rfc822: add Origin.relative_to() method
The method returns a new Origin class with source component altered to
be relative to the specified base_dir argument. This is useful for
displaying short origin specifications when the typical source contains
an absolute path to some file but the user is "close" to that file and
expects relative filenames to be displayed.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>