scripts/sleep_test_log_check: Rewrote the parser logic.
It's now much more strict in how it finds stuff and will readily
complain if it expected a certain format and markers in the file but
didn't find them.

This also upgrades the parser to handle the fwts 13.x format, may be
incompatible with older versions of fwts which are not recommended for
use with checkbox anyway.