plainbox:secure:config: add a hack that elevates NotUnsetValidator
This hack ensures that the internal KindValidator is added _after_
the NotUnsetValidator, if the latter is present as the first item
in the list of custom validators.

This is meant to give better user experience when handling unset
variables, instead of a confusing message that type does not match
it allows the NotUnsetValidator to provider a more appropriate message
stating that the variable is simply not set.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>