plainbox:secure:rfc822: fix quoting of leading dot
This patch fixes the behavior for quoting the lading dot on an otherwise
empty line of a multi-line value of RFC822 record entries. The
replacement syntax allows arbitrary text after the special value
'<space>.' (which is stripped-away). This retains compatibility with the
previous parser and makes it generic, up to any "depth" of dots.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>