checkbox-old:scripts: New screenshot_validation test using OpenCV
This new test validates a webcam screenshot using OpenCV ORB detection and
a FLANN Matcher (Fast Approximate Nearest Neighbor Search Library).

The check operates on GRAYSCALE images, so both the captured image and the
query image are converted to this format.

The test supports different level of ambient light and is quite robust to
distortions (lid does not need to be perfectly vertical)

For better results, the use of HD webcams is recommended (min. 1600x900)