support: detect broken virtualbox setups
On some cloud images that we are using the host-guest filesystem is not
working. When accessed from within the guest it causes processes to be
killed after a brief timeout. For now this affects saucy (but it is
likely to come back again in Trusty+1 since it's a common issue).

This patch adds a simple check to see if we can really use the /vagrant
directory where that filesystem is mounted and provides a helpful
message that may offer a work-around. The work around requires a little
bit of setup on the host side so it cannot be provided automatically.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>