checkbox-gui: Add TRANSLATORS comments and fix welcome text to work with xgettext.
The welcome message text in WelcomeView.qml was a multi-line string - xgettext
can't handle this so we need to change it to be one line.

We can also use comments beginning with the word TRANSLATORS to pass notes to
the translators to make sure they understand the text being translated. We use
this in a few places to make sure things such as HTML tags are not translated,
and to clear up ambiguity.

Signed-off-by: Brendan Donegan <brendan.donegan@canonical.com>