checkbox_ng:commands:cli: New SelectableJobTreeNode class
Implementation of a node in a tree that can be selected/deselected.
Categories can be expanded/collapsed.

This class offers a convenient method render() which returns the tree as a
simple list of categories and jobs suitable for display. Jobs are properly
indented to respect the tree hierarchy and selection marks are added
automatically at the beginning of each element.