plainbox:session: add SessionManager.create_with_state()
This new method allows to construct a SessionManager (which couples
session state and storage) from an existing state instance. It is useful
when a program manipulates the state object and needs to execute jobs
(which require storage to be available) without having to import a
number of other helper classes.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>