checkbox-old:jobs: Update job descriptions for USB tests.
Updated a lot of the USB jobs to make them more suitable for use in the
IHV USB test suite. Make it clear that HDDs are preferred for testing
and remove the local job for usb3 performance and replace it with a single
job. Also add a usb 2.0 performance job. Add estimated durations as well.

Updated client-cert and client-selftest whitelists with the new usb3
performance job.

Signed-off-by: Brendan Donegan <brendan.donegan@canonical.com>