plainbox:providers:special: adjust providers after API changes
This patch adjusts CheckBoxSrcProvider to the new Provider1 API. As a
small tweak, the description field of both providers is now marked with
N_ (no-op version of gettext)

StubBoxProvider is a little bit special, as we need to share its
meta-data with manage.py so instead of making it another provider
subclass, we just provider a new function, get_stubbox_def(), that
returns a Provider1Definition for StubBox. All existing call sites are
patched to use the new interface instead.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>