plainbox:secure:providers: add Provider1.namespace
This patch adds a namespace property to each provider. The namespace is
defined as the name, all the way up to the colon. This allows
organizations to split any set of jobs into multiple providers but still
use jobs from all the providers inside the same namespace without any
additional syntax.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>