support: make test-in-coverage reuse $VIRTUAL_ENV
The full coverage report, as of this writing, is around 11 megabytes
big. Writing that on slower disks (lots of small files) takes a
considerable amount of time. To make testing faster, we can reuse the
$VIRTUAL_ENV variable, which _may_ be pointing to a directory like
/ramdisk which _might_ be tmpfs-based, which is definitely faster than
any normal filesystem and also lessens the wear on SSDs.
This patch implements exactly that behavior.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>