plainbox:job: add JobDefinition.partial_id
This new property is *exactly* as JobDefinition.id used to be, the plain
identifier of a job, without any name-space support. It should not be
used, except in special cases for display as it will otherwise lead to
clashes if two or more providers contain a job with the same partial_id.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>