checkbox-gui: Implement whitelist ordering
Refactor the logic used by checkbox-gui to select which jobs to run.
Instead of manually cross-referencing the jobs available with the
whitelists selected, use SelectJobs to do the same thing, but guarantee
the jobs are returned in whitelist order. Also slightly tweak some bits of
code so that this order is always retained throughout the process, up until
the final run list is generated.

Signed-off-by: Brendan Donegan <brendan.donegan@canonical.com>