plainbox:secure:rfc822: additionally store raw rfc822 data
This patch extends the RFC822 parser to store both the normalized data
(as before) and the raw data which may include extra spaces and the
magic dot syntax.

The resulting raw data is identical to those obtained by xgettext,
except is has a trailing newline that needs to be removed.

Some examples of raw data:

key: value # (where the newline is after the pound or hash sign)

Is parsed as: {'key': ' value '} whereas in the normalized copy both the
leading and trailing whitespace are lost.

key:
  value

Is parsed as: {'key': ' value'} (mind the fact that one space was
discarded but the other one remained. In the normalized copy all leading
spaces of consistent length on all the lines are lost.

key:
 .
 value

Is parsed as {'key': '.\nvalue'} whereas in the normalized copy the dot
is stripped away.

Lastly trailing newlines (if any) are retained. Let's assume that
the following example has a newline at the end (and all the examples
above didn't)

key:
 value\n

Is parsed as {'key': 'value\n'}.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>