plainbox,checkbox: make CheckBoxInvocationMixIn pass config object around
This patch sets the stage for the next patch which will rely on having
the configuration object available in several commands. To make that
possible the generic invocation mix-in is amended to require two arguments.

This puts code all around the tree and again requires locks-step update
of checkbox and plainbox.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>