providers:resource: Added requirements resource
The requirements resource reads a file in the provider's data directory
which provides a URL to provider-relevant data, this can later be shown
in the HTML report and will be included in the submission XML as well.

The resource was also added to the submission-resources metajob in the
checkbox provider.