checkbox-gui: Added support for client-name as an option to exporters.
when saving both html and xml, the applicationName will be sent as an
option to the exporters, to be used as the client-name that will appear
in the reports.

Note I didn't use the existing applicationName variable because it has
special meaning to QML to configure storage paths, so it can't be just
renamed (but also doesn't just work if used), so I simply copied the
value on another variable.