plainbox-provider-checkbox: Update default.whitelist to allow LP submission
This change is meant to support the LaunchpadTransport which is coming in
another merge request, however it can stand on its own. We need to add the
'dmi' resource to the miscellanea/submission-resources job, this is
required to generate 'unique' fingerprints for the submission form. We
also need to use miscellanea/submission-resources in default.whitelist,
so that the necessary resource jobs will be run (checkbox-gui does not
run resource jobs unless another job depends on them, as per
https://bugs.launchpad.net/bugs/1302388).

Signed-off-by: Brendan Donegan <brendan.donegan@canonical.com>