checkbox-support:parsers:udevadm: Updated heuristic for some VIDEO devices.
Usually a video device has a PCI subclass_id of Pci.CLASS_DISPLAY_VGA or
Pci.CLASS_DISPLAY_3d. However, some manufacturers which hadn't
previously used a subclass_id of Pci.CLASS_DISPLAY_OTHER (0x80 or
decimal 128) have begun to do so (as of 2013-2014); and others which
previously used this class for uninteresting devices are now using it
for actual video devices.

Previously we identified AMD as doing this, and added a simple heuristic
to accomodate that; on a few newer systems, Intel has also started using
the OTHER display class, but erratically. This code implements a more
versatile heuristic we can expand if/when we find other vendors/devices
that break our old expectations.

A previous commit added a test and data file to confirm that things
behave as expected, and also that none of the other tests break,
indicating that this change doesn't cause old systems to "sprout" new
VIDEO devices on account of faulty heuristics.