Mount root by label
Instead of keeping a map of qemu disk types to /dev files, we can just mount by
label and let the initramfs detect which device file to use. This avoids having
to deal with differences between partitioned disks and unpartitioned disks,
as well as not having to predict the order in which the seed and root disks
get enumerated.