plainbox:commands:run: change multiple-session resume error paths
This patch changes how we resume if there are multiple possible
candidates. First of all, the user can now choose the session to resume.
A session may fail to resume and this is now handled. Lastly, after we
decline to resume any session we prompt the user to create a new one.

This is intended to prevent the user from hoping to resume, see the resume
fail and then proceed to run something without a chance to abort.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>