providers:cert-client: add more LED tests to client-cert whitelist
This adds almost all of the LED tests to the client-cert.whitelist. The
excluded jobs include all of the suspend/led_after_suspend/.* tests (per
Ara's request) as well as the led/wireless which is redundant given that
we have led/bluetooth and led/wlan in the same group already.

Fixes: https://bugs.launchpad.net/plainbox-provider-canonical-certification/+bug/1316519

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>