plainbox:ctrl: add CheckBoxExecutionController.get_warm_up_for_job
This patch adds a dummy (no-op) but correct implementation of the
new interface for execution controllers that allows them to provide
a callable needed to "warm-up" the system to run a particular job.

This interface is designed to couple warm up implementation with
particular execution controller classes. Currently only one controller
actually supports warm-up but none of the upper layers are aware of that
and they call pointless warm-up entirely needlessly.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>