plainbox:runner: allow to customize execution controllers
This patch adds a new optional argument to JobRunner that allows one to
pass an explicit list of execution controllers that the runner will have
access to. This can prevent usage of the default list of execution
controllers and specifically to prevent them from initializing in cases
where that may be undesired.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>